+++ src/main/java/com/portfordev/pro/controller/main_controller.java
@@ -2,9 +2,7 @@ package com.portfordev.pro.controller;
 
 import java.io.File;
 import java.io.FilenameFilter;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
