+++ src/main/webapp/resources/css/alert_custom.css
@@ -1,0 +1,88 @@
+@charset "UTF-8";
+/* 
+	단위
+	em : 최상위 속성에 값을 상속받아 %단위로 비례함 1.2 -> 120%
+	rem : em하위에 em이 중첩되면 중첩이 누적되기 때문에 root로부터 계산할 수 있도록 함(일반적으로 html태그에 붙이는듯)
+	vh, vw : fixed 조건에서 각 높,너비를 1/100로 표현
+	vmin, vmax : 화면의 높,너비 값 중 min과 max
+*/
+/*
+	main theme color : #5385c1
+*/
+/* 안내문구를 보여줄 section */
+#alert-wrap{
+	position: fixed;
+    display: none;
+    background: rgba(0,0,0,0.5);
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    width: 100%;
+    height:100%;
+    z-index: 999;
+}
+#alert-wrap .alert-box{
+	position: absolute;
+    background: #fff;
+    border-radius: 6px;
+    width: 400px;
+    max-width:90%;
+    top: calc(50% - 100px);
+    left: calc(50% - 200px);
+    text-align: left;
+}
+.alert-box img{
+    height: 22px;
+    margin-left: 17px;
+    margin-top: 17px;
+    margin-right: 0;
+    vertical-align: top;
+}
+.alert-box b{
+	display:inline-block;
+	padding:5px;
+	font-size:1.2em;
+	text-align:left;
+	margin-top:10px;
+	margin-left:0;
+}
+.alert-box hr{
+	margin: 1em 0;
+}
+.alert-box strong{
+	font-weight:bold;
+	color:black;
+}
+.alert-box .alert-message{
+	display:none;
+	white-space:normal;
+	word-break:break-all;
+	padding:10px;
+	text-align:center;
+}
+.site-alert .site-url{
+	display:block;
+	white-space:normal;
+	word-break:break-all;
+	min-height: 30px;
+}
+#alert-wrap .alert-box .alert-btn-area{
+	position:relative;
+	bottom:0;
+	text-align: right;
+}
+.alert-btn-area button{
+	margin-right:10px;
+	margin-bottom:10px;
+	background:#5385c1;
+	outline:none!important;
+	cursor:pointer;
+	border-radius:4px;
+}
+.alert-btn-area .goBtn{
+	color:#fff;
+}
+.alert-btn-area .cancelBtn{
+	background:transparent;
+}
