+++ src/main/webapp/resources/css/portfolio_collection.css
@@ -80,6 +80,12 @@ tbody{
 	padding: 11px 16px;
 	margin-right:25px;
 }
+.category-list-item>img{
+	height: 19px;
+    width: auto;
+    margin-right: 5px;
+    vertical-align: middle;
+}
 .category-list .active-category{
 	color:#5385c1;
 	font-size:16.5px;
@@ -96,18 +102,24 @@ tbody{
 }
 @media screen and (max-width: 1024px){
 	.category-list .category-list-item{
-	
+	    margin-right: 10px;
+	    padding-right: 10px;
 	}
 }
 @media screen and (max-width: 768px){
 	.category-list .category-list-item{
-	
+		margin-right: 0px;
+	    padding-right: 5px;
 	}
 }
 
-@media screen and (max-width: 510px){
+@media screen and (max-width: 548px){
 	.category-list{
 		width:100%;
+	}
+	.category-list .category-list-item{
+		margin-right: 0px;
+	    padding-right: 0px;
 	}
 	
 }
