+++ src/main/webapp/resources/css/portfolio_manager.css
@@ -1,0 +1,203 @@
+@charset "UTF-8";
+.not-scroll{
+	overflow:hidden!important;
+	touch-action:none!important;
+}
+#portfolio-manager{
+	position:fixed;
+	top:0;
+	left:0;
+	right:0;
+	bottom:0;
+	width:100%;
+	height:100%;
+    background: rgba(0,0,0,0.8);
+    display:none;
+	z-index: 105;
+	overflow: hidden;
+}
+#exit-manager{
+	position:absolute;
+	top:50px;
+	right:50px;
+	font-size:40px;
+	font-weight:bold;
+	color:white;
+	z-index:199;
+}
+#exit-manager:hover{
+	cursor:pointer;
+	opacity:0.8;
+	transition:0.3s;
+}
+#portfolio-manager-container{
+	width:calc(100% - 200px);
+	height:calc(100% - 100px);
+	position:absolute;
+	left:100px;
+	right:100px;
+	top:100px;
+	bottom:0;
+	padding-top:10px;
+	padding-bottom:0;
+	z-index:106;
+}
+#portfolio-add-area{
+	display:block;
+    height: 42px;
+    min-width:455px;
+}
+#write-portfolio{
+	background-color: white!important;
+	display:inline-block;
+    border-radius: 5px 5px 0px 0px;
+	float:right;
+    height: 42px;
+}
+#write-portfolio:hover{
+	opacity:0.9;
+    border-radius: 10px 10px 0px 0px;
+	transition:0.4s;
+}
+#write-portfolio>img{
+    vertical-align: sub;
+}
+.loading-myList{
+    height: 303px;
+    padding: 5px;
+    text-align: center;
+    margin-top:5px;
+    margin-bottom: 25px;
+}
+.loading-myList:hover{
+	cursor:wait;
+}
+.loading-myList>img{
+	height:100%;
+	width:auto;
+	margin:0 auto;
+}
+#my-portfolio-list{
+	color: white;
+    background: white;
+    max-height: calc(100% - 52px);
+    min-width: 455px;
+    max-width: 100%;
+    min-height: 455px;
+    margin: 0 auto;
+    display: block;
+    text-align: center;
+    padding: 65px 10px;
+    border-radius: 5px 0px 5px 5px;
+    overflow-y: scroll;
+    /* scroll bar 감주기 */
+	-ms-overflow-style:none;	/* IE and Edge */
+	scrollbar-width:none;	/* Firefox */
+}
+#my-portfolio-list::-webkit-scrollbar{
+	display:none;	/* Chrome, Safari, Opera */
+}
+.no-portfolio-item{
+    text-align: center;
+    font-weight: bold;
+    font-size: 22px;
+    margin-top: 40px;
+    z-index: 999;
+    background: #FFF;
+    color: #888;
+    display: block;
+}
+.my-portfolio-item{
+    height: 283px;
+    display: inline-block;
+    width: calc(40% - 54px);
+    min-width: 400px;
+    max-width: 450px;
+    margin: 17px 17px;
+    background: #FFF;
+    box-shadow: 1px 1px 4px rgba(0,0,0,0.2);
+    position: relative;
+    color: rgb(0, 0, 0);
+}
+.my-portfolio-item:hover{
+	cursor:pointer;
+	border-radius:2%;
+	transition:0.3s;
+	opacity:0.7;
+}
+.mp-img-wrapper{
+	width: 100%;
+	height:283px;
+    max-height: 283px;
+    background-size: cover;
+    background-position: top;
+    background-repeat: no-repeat;
+    background-attachment: scroll;
+    text-align: center;
+    border-top: 0.5em solid #5385c1;
+    box-shadow: 1px 1px 4px rgba(0,0,0,0.2);
+}
+.mp-info-wrapper{
+	position:absolute;
+	display:none;
+	width:100%!important;
+	height:100%!important;
+	background:rgba(255, 255, 255, 0.7);
+	left:0;
+	right:0;
+	bottom:0;
+	top:0;
+	padding-top:15%;
+}
+.mp-info-wrapper img{
+	display:inline-block!important;
+}
+.mp-info-subject{
+	font-weight:bold;
+	font-size:1.5em;
+	display:block;
+	text-align:center;
+    margin: 1em auto;
+    vertical-align:top;
+}
+.mp-icon{
+	height:30px;
+	width:auto;
+    margin: 0 5px;
+}
+.mp-count{
+	height:40px;
+	font-size:1em;
+    margin-right: 5px;
+    font-size: 18pt;
+    line-height: 18pt;
+    vertical-align:top;
+}
+.mp-controller form{
+	margin: 0 10px;
+    text-align: center;
+    display: inline-block;
+    width: auto;
+    height:auto;
+}
+.mp-controller button{
+	margin-top: 20px;
+    border-radius: 5px;
+    background: white;
+    display:inline-block;
+    outline:none!important;
+}
+.mp-controller button:hover{
+	background:#999;
+	color:#FFF;
+}
+/* 모바일 */
+@media screen and (max-width:650px){
+	
+#portfolio-manager-container{
+	width:calc(100% - 40px);
+	left:20px;
+	right:20px;
+	padding-top:10px;
+	padding-bottom:0;
+}
