+++ src/main/webapp/resources/css/profile_main_slidebar.css
@@ -1,7 +1,7 @@
 .slide{color:white;max-width:100%;margin:0 auto;margin-top:10px;}
 .slick-track{width:100%}
    .slide_in{height:400px;margin: 10px;background:white;vertical-align: center;min-width:100px}
-   .slide_in_content{width:100%;height:400px; filter: brightness(70%); opacity: 0.6}
+   .slide_in_content{width:100%;height:400px;opacity: 1}
    .right{display: inline-block;}
    .left{display: inline-block;}
    .h1{font-color:black;text-align: center;font-size: 30pt;font-weight:700;color:gray}
